Options model
* also fixed a bug that files only get deleted if present
  as PStore seems to only create a file if something is
  actually saved (which makes sense)

+++ spec/pomodoro_tracker/file_persistor_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe 'Persistor' do
 
-  TEST_FILE_PATH = File.expand_path('../', __FILE__) + '/test_storage.pstore'
+  TEST_FILE_PATH = File.expand_path('../', __FILE__) + '/test_storage.yml'
 
   before :each do
     @activity = FactoryGirl.build :activity
@@ -10,7 +10,7 @@ describe 'Persistor' do
   end
 
   after :each do
-    File.delete TEST_FILE_PATH
+    File.delete TEST_FILE_PATH if File.exist? TEST_FILE_PATH
   end
 
   it 'saves the activity and returns a truthy value' do
